+++ lib/src/boxy/box_delegate.dart
@@ -101,28 +101,45 @@ class RenderBoxy<ChildHandleType extends BaseBoxyChild> extends RenderBox
   double computeMinIntrinsicWidth(double height) {
     updateChildHandles();
     return wrapPhase(
-        BoxyDelegatePhase.intrinsics, () => delegate.minIntrinsicWidth(height));
+      BoxyDelegatePhase.intrinsics,
+      () => delegate.minIntrinsicWidth(height),
+    );
   }
 
   @override
   double computeMaxIntrinsicWidth(double height) {
     updateChildHandles();
     return wrapPhase(
-        BoxyDelegatePhase.intrinsics, () => delegate.maxIntrinsicWidth(height));
+      BoxyDelegatePhase.intrinsics,
+      () => delegate.maxIntrinsicWidth(height),
+    );
   }
 
   @override
   double computeMinIntrinsicHeight(double width) {
     updateChildHandles();
     return wrapPhase(
-        BoxyDelegatePhase.intrinsics, () => delegate.minIntrinsicHeight(width));
+      BoxyDelegatePhase.intrinsics,
+      () => delegate.minIntrinsicHeight(width),
+    );
   }
 
   @override
   double computeMaxIntrinsicHeight(double width) {
     updateChildHandles();
     return wrapPhase(
-        BoxyDelegatePhase.intrinsics, () => delegate.maxIntrinsicHeight(width));
+      BoxyDelegatePhase.intrinsics,
+      () => delegate.maxIntrinsicHeight(width),
+    );
+  }
+
+  @override
+  double? computeDistanceToActualBaseline(TextBaseline baseline) {
+    updateChildHandles();
+    return wrapPhase(
+      BoxyDelegatePhase.intrinsics,
+      () => delegate.distanceToBaseline(baseline),
+    );
   }
 
   @override
@@ -182,6 +199,11 @@ class RenderBoxy<ChildHandleType extends BaseBoxyChild> extends RenderBox
       },
     );
   }
+
+  @override
+  void handleEvent(PointerEvent event, BoxHitTestEntry entry) {
+    return _delegate.onPointerEvent(event, entry);
+  }
 }
 
 /// Mixin for the logic shared by [BoxBoxyDelegate] and [BoxyDelegate].
@@ -226,6 +248,9 @@ mixin BoxBoxyDelegateMixin<LayoutData extends Object,
   /// actual orientation. Additionally, the [inflate] method will throw
   /// [CannotInflateError] if called during a dry layout.
   Size layout() => constraints.smallest;
+
+  @override
+  Size get renderSize => render.size;
 
   /// Override to change the minimum width that this box could be without
   /// failing to correctly paint its contents within itself, without clipping.
@@ -299,6 +324,19 @@ mixin BoxBoxyDelegateMixin<LayoutData extends Object,
     return 0.0;
   }
 
+  /// Override to return the distance from the y-coordinate of the position of
+  /// the box to the y-coordinate of the first given baseline in the box's
+  /// contents, if any, or null otherwise.
+  ///
+  /// See also:
+  ///
+  ///  * [RenderBox.computeDistanceToActualBaseline], which has usage examples.
+  double? distanceToBaseline(TextBaseline baseline) => null;
+
+  @override
+  void onPointerEvent(PointerEvent event, BoxHitTestEntry entry) =>
+      super.onPointerEvent(event, entry);
+
   @override
   BoxHitTestResult get hitTestResult => super.hitTestResult as BoxHitTestResult;
 
@@ -322,7 +360,7 @@ mixin BoxBoxyDelegateMixin<LayoutData extends Object,
 ///  * [BoxyLayerContext], a wrapper that can push [Layer]s.
 ///  * [BoxyDelegate], a base delegate that supports both [BoxyChild] and
 ///    [SliverBoxyChild].
-class BoxBoxyDelegate<LayoutData extends Object>
+abstract class BoxBoxyDelegate<LayoutData extends Object>
     extends BaseBoxyDelegate<LayoutData, BaseBoxyChild>
     with BoxBoxyDelegateMixin<LayoutData, BaseBoxyChild> {
   /// Constructs a BoxyDelegate with optional [relayout] and [repaint]
@@ -413,7 +451,7 @@ class BoxBoxyDelegate<LayoutData extends Object>
 ///   @override
 ///   void paint() {
 ///     canvas.drawRect(
-///       Offset.zero & render.size,
+///       Offset.zero & renderSize,
 ///       Paint()..color = Colors.blue,
 ///     );
 ///   }