+++ lib/src/boxy_flex.dart
@@ -1,4 +1,5 @@
 import 'dart:math' as math;
+import 'dart:math';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -12,10 +13,14 @@ enum BoxyFlexIntrinsicsBehavior {
   /// Measure the intrinsic main axis of inflexible children with an infinite
   /// max cross axis size, using it as the max main axis size of the dominant
   /// child.
+  ///
+  /// TL;DR: Use this if you have text in a horizontal Flex.
   measureMain,
 
   /// Measure the intrinsic cross axis of the dominant child with an infinite
   /// max main axis size.
+  ///
+  /// TL;DR: Use this if you want speed, since it only measures one child.
   measureCross,
 }
 
@@ -1119,6 +1124,9 @@ class RenderBoxyFlex extends RenderBox
     required ChildLayouter layoutChild,
   }) {
     int totalFlex = 0;
+    final double minMainSize = _direction == Axis.horizontal
+        ? constraints.minWidth
+        : constraints.minHeight;
     final double maxMainSize = _direction == Axis.horizontal
         ? constraints.maxWidth
         : constraints.maxHeight;
@@ -1417,9 +1425,12 @@ class RenderBoxyFlex extends RenderBox
     }
 
     return _LayoutSizes(
-      mainSize: canFlex && mainAxisSize == MainAxisSize.max
-          ? maxMainSize
-          : allocatedSize,
+      mainSize: max(
+        minMainSize,
+        canFlex && mainAxisSize == MainAxisSize.max
+            ? maxMainSize
+            : min(maxMainSize, allocatedSize),
+      ),
       crossSize: crossSize,
       allocatedSize: allocatedSize,
       needsBaseline: needsBaseline,